Fix gif url
When README is imported to developer portal, the relative urls are fixed to point to the git repository using `awk` and `sed` scripts.

For the images it point to the raw.github.com url, but the url of the image have to be just after the image declaration on the markdown

```
![txt](./relative/path.jpg)
```

However, for the ESM example gif, the format was different, causing that the fixed link doesn't point to the `raw.` url. It was:

```
![text]
//...

[text]: ./relative/path.jpg
``` 

So it causes that replaced url is to `github.com` causing image break on the developer portal.